refactor: extract util functions for working with .artilleryrc
