+++ packages/artillery/lib/telemetry.js
@@ -7,7 +7,7 @@
 const { version: artilleryVersion } = require('../package.json');
 const { isCI, name: ciName } = require('ci-info');
 const chalk = require('chalk');
-const { readArtilleryConfig, updateArtilleryConfig } = require('./util');
+const { readArtilleryConfig, updateArtilleryConfig } = require('./utils-config');
 const debug = require('debug')('telemetry');
 
 const POSTHOG_TOKEN = '_uzX-_WJoVmE_tsLvu0OFD2tpd0HGz72D5sU1zM2hbs';
